+++ CHANGES.md
@@ -73,6 +73,7 @@ To be released.
     parameter.  This purposes to determine how long to wait for pending
     messages when a `Swarm` instance is requested to terminate.
  -  Added `NamespaceNotFoundException` class.  [[#232]]
+ -  Added `SimultaneousTxsException` class.  [[#242]]
 
 ### Behavioral changes
 
@@ -94,6 +95,8 @@ To be released.
  -  `Transaction<T>.EvaluateActionsGradually()` became to record
     `IAccountStateDelta.SetState()` calls even if its argument is the same
     to the previous state. [[#241]]
+ -  A signer became to allowed to have only one transaction at most in 
+    `Block<T>.Transactions`. [[#125]], [[#242]]
 
 ### Bug fixes
 
@@ -125,6 +128,7 @@ To be released.
 
 [Ethereum Homestead algorithm]: https://github.com/ethereum/EIPs/blob/master/EIPS/eip-2.md
 [#31]: https://github.com/planetarium/libplanet/issues/31
+[#125]: https://github.com/planetarium/libplanet/issues/125
 [#185]: https://github.com/planetarium/libplanet/pull/185
 [#187]: https://github.com/planetarium/libplanet/issues/187
 [#190]: https://github.com/planetarium/libplanet/pull/190
@@ -151,6 +155,7 @@ To be released.
 [#236]: https://github.com/planetarium/libplanet/pull/236
 [#240]: https://github.com/planetarium/libplanet/pull/240
 [#241]: https://github.com/planetarium/libplanet/pull/241
+[#242]: https://github.com/planetarium/libplanet/pull/242
 
 
 Version 0.2.2